# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: assignments
    description: Assignment operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List all assignments
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create a new assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentBody"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"

  /assignments/summary:
    get:
      operationId: getAssignmentsSummary
      tags: [assignments]
      summary: Get summary stats
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentsSummary"

  /assignments/{id}:
    get:
      operationId: getAssignment
      tags: [assignments]
      summary: Get an assignment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "404":
          description: Not found
    patch:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update an assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssignmentBody"
      responses:
        "200":
          description: Assignment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "404":
          description: Not found
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Delete an assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Assignment:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        course:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, done]
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - course
        - dueDate
        - status
        - priority
        - createdAt

    CreateAssignmentBody:
      type: object
      properties:
        title:
          type: string
        course:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
        priority:
          type: string
          enum: [low, medium, high]
      required:
        - title
        - course
        - dueDate
        - priority

    UpdateAssignmentBody:
      type: object
      properties:
        title:
          type: string
        course:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, done]
        priority:
          type: string
          enum: [low, medium, high]

    AssignmentsSummary:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        done:
          type: integer
        overdue:
          type: integer
        dueSoon:
          type: integer
      required:
        - total
        - pending
        - done
        - overdue
        - dueSoon