# lib/api-zod/src/generated/types/assignment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssignmentPriority } from "./assignmentPriority";
import type { AssignmentStatus } from "./assignmentStatus";

export interface Assignment {
  id: number;
  title: string;
  course: string;
  description?: string | null;
  dueDate: Date;
  status: AssignmentStatus;
  priority: AssignmentPriority;
  createdAt: Date;
}