# lib/api-zod/src/generated/types/assignmentsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AssignmentsSummary {
  total: number;
  pending: number;
  done: number;
  overdue: number;
  dueSoon: number;
}