# lib/api-zod/src/generated/types/createAssignmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAssignmentBodyPriority } from "./createAssignmentBodyPriority";

export interface CreateAssignmentBody {
  title: string;
  course: string;
  description?: string | null;
  dueDate: Date;
  priority: CreateAssignmentBodyPriority;
}