# lib/api-zod/src/generated/types/updateAssignmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAssignmentBodyPriority } from "./updateAssignmentBodyPriority";
import type { UpdateAssignmentBodyStatus } from "./updateAssignmentBodyStatus";

export interface UpdateAssignmentBody {
  title?: string;
  course?: string;
  description?: string | null;
  dueDate?: Date;
  status?: UpdateAssignmentBodyStatus;
  priority?: UpdateAssignmentBodyPriority;
}